[Refactor] response 패키지 생성
